pyproject.toml
[tool.poetry]
name = "bybitp2pbot"
version = "0.1.0"
description = ""
authors = ["lionex-ui <eugenemescaline@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^3.8.0"
betterconf = "^2.8.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.31"
alembic = "^1.13.1"
psycopg2 = "^2.9.9"
asyncpg = "^0.29.0"
aioredis = "^2.0.1"
aiosqlite = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
